🧪 Fix WaitTimeManagerImplTest and enhance spot matching logic
- Fix type inference errors in test assertions by specifying Double types
- Update test expectations to match absolute min/max limit behavior
- Implement distance-based spot matching instead of exact equality comparison
- Add test state cleanup to prevent test interference
- Fix PlayerFishingListener compilation error with type casting
- Add clearAllEffects method for test isolation

All WaitTimeManagerImplTest cases now pass successfully.

+++ api/src/main/kotlin/party/morino/moripafishing/api/core/fishing/WaitTimeManager.kt
@@ -51,5 +51,5 @@ interface WaitTimeManager {
      * @param angler プレイヤー
      * @return ルアータイム(秒)
      */
-    fun getWaitTime(angler: Angler): Pair<Int, Int>
+    fun getWaitTime(angler: Angler): Pair<Double, Double>
 }
